 \docType{data}
    \name{varambally}
    \alias{varambally}
    \title{The Cai Cohort}
    \format{a \code{ExpressionSet} instance, 1 row per Affymetrix microarray probe.}
    \source{
      GSE32269
    }
    \description{
      Object imported using GEOquery and curated to remove variables introduced by GEO
    }
    \author{
Mark Dunning
    }
    \usage{varambally}
    \keyword{datasets}